Moved spotifyapi objects inside express middleware + ...
...added request token condition to users/login axios request

+++ frontend/src/App.js
@@ -70,36 +70,37 @@ function App() {
 
   //User Effect hook for logging in the user with API upon refreshToken update
   useEffect(() => {
-    //Start loading
-    setIsLoading(true);
+    if (refreshToken) {
+      //Start loading
+      setIsLoading(true);
+      axios
+        .post(process.env.REACT_APP_BACKEND_URL + "/users/login", {
+          refreshToken: refreshToken,
+        })
+        .then((data) => {
+          //if login is unsuccessful, reset refresh token
+          if (data.data.error) {
+            dispatch(updateRefreshToken(""));
+            localStorage.clear();
+            setFaultyLogin(true);
+          }
+          //if login is sucessful, update user object
+          else {
+            dispatch(updateUser(data.data.return));
+            setFaultyLogin(false);
+          }
 
-    axios
-      .post(process.env.REACT_APP_BACKEND_URL + "/users/login", {
-        refreshToken: refreshToken,
-      })
-      .then((data) => {
-        //if login is unsuccessful, reset refresh token
-        if (data.data.error) {
-          dispatch(updateRefreshToken(""));
-          localStorage.clear();
-          setFaultyLogin(true);
-        }
-        //if login is sucessful, update user object
-        else {
-          dispatch(updateUser(data.data.return));
-          setFaultyLogin(false);
-        }
+          //Stop loading
+          setIsLoading(false);
+        })
+        .catch((err) => {
+          //Stop loading
+          setIsLoading(false);
 
-        //Stop loading
-        setIsLoading(false);
-      })
-      .catch((err) => {
-        //Stop loading
-        setIsLoading(false);
-
-        console.log(err);
-      });
-  }, [refreshToken]);
+          console.log(err);
+        });
+    }
+  }, [refreshToken, dispatch]);
 
   //User Effect to get user group list when userObject is updated
   useEffect(() => {
